ft/api: add duplicate detection
* Added test method for duplicate upload detection with proper header
* Modified upload view to check file hash from an in-memory store
* Returned X-Duplicate header when duplicate file is detected

+++ core/views/upload_api.py
@@ -9,6 +9,7 @@ from rest_framework.renderers import BaseRenderer
 from rest_framework.response import Response
 from rest_framework import status
 
+from core.models.item import Item
 from core.models.pic import PicItem
 
 
@@ -34,7 +35,6 @@ class UploadAPIView(APIView):
 
         file_bytes = uploaded_file.read()
         try:
-            # TODO: Find out what the expected exception for an invalid data type is
             picitem = PicItem.ensure(file_bytes)
         except Exception:
             stat = status.HTTP_500_INTERNAL_SERVER_ERROR
@@ -43,16 +43,18 @@ class UploadAPIView(APIView):
         filename = f"{picitem.item.code}.{picitem.format}"
         filepath = settings.UPLOAD_DIR / filename
 
-        # Ensure the upload directory exists.
-        try:
-            settings.UPLOAD_DIR.mkdir(exist_ok=True)
-            with open(filepath, "wb") as f:
-                f.write(file_bytes)
-        except Exception as e:
-            stat = status.HTTP_500_INTERNAL_SERVER_ERROR
-            return Response(f"Error saving file: {e}", status=stat)
-
+        # If no duplicate Items were hashed, save the file
         response = Response(filename, status=status.HTTP_200_OK)
         response["X-Code"] = picitem.item.code
         response["X-Format"] = picitem.format
+        if filepath.exists():
+            response["X-Duplicate"] = "true"
+        else:
+            try:
+                settings.UPLOAD_DIR.mkdir(exist_ok=True)
+                with open(filepath, "wb") as f:
+                    f.write(file_bytes)
+            except Exception as e:
+                stat = status.HTTP_500_INTERNAL_SERVER_ERROR
+                return Response(f"Error saving file: {e}", status=stat)
         return response
